requirements.txt
scikit-learn>=1.0.0
pandas>=2.0.0
numpy>=1.21.0
pytest>=7.0.0
black>=22.0.0
jupyter>=1.0.0
matplotlib>=3.0.0
PyYAML>=6.0
streamlit>=1.0
joblib>=1.0